#26 예비 착한 밥상 조회 & 제보 & 좋아요 클릭 api 구현

+++ app/src/main/java/com/hackathon/alddeul_babsang/domain/repository/ReportRepository.kt
@@ -4,7 +4,7 @@ import com.hackathon.alddeul_babsang.data.dto.response.ResponseReportDto
 
 interface ReportRepository {
     suspend fun postReports(
-        userId: Long
+        userId: Int
     ): Result<List<ResponseReportDto>>
 
     suspend fun postReportWrite(
